+++ src/sock/fake_ipv4.go
@@ -22,8 +22,12 @@ func BuildFakeSNIPacketV4(original []byte, cfg *config.SetConfig) []byte {
 		rand.Read(fakePayload)
 	case config.FakePayloadCustom:
 		fakePayload = []byte(cfg.Faking.CustomPayload)
+	case config.FakePayloadDefault1:
+		fakePayload = FakeSNI1
+	case config.FakePayloadDefault2:
+		fakePayload = FakeSNI2
 	default:
-		fakePayload = FakeSNI
+		fakePayload = FakeSNI1
 	}
 
 	fakeLen := ipHdrLen + tcpHdrLen + len(fakePayload)